+++ base/util/src/com/netscape/cmsutil/util/Utils.java
@@ -98,7 +98,7 @@ public class Utils {
         ByteArrayOutputStream out = new ByteArrayOutputStream(s.length());
 
         for (int i = 0; i < s.length(); i++) {
-            int c = (int) s.charAt(i);
+            int c = s.charAt(i);
 
             if (c == '+') {
                 out.write(' ');
@@ -120,7 +120,7 @@ public class Utils {
         ByteArrayOutputStream out = new ByteArrayOutputStream(s.length());
 
         for (int i = 0; i < s.length(); i++) {
-            int c = (int) s.charAt(i);
+            int c = s.charAt(i);
 
             if (c == '+') {
                 out.write(' ');
