+++ base/util/src/netscape/security/x509/PrivateKeyUsageExtension.java
@@ -149,7 +149,7 @@ public class PrivateKeyUsageExtension extends Extension
         for (int i = 0; i < seq.length; i++) {
             DerValue opt = seq[i];
 
-            if (opt.isContextSpecific((byte) TAG_BEFORE) &&
+            if (opt.isContextSpecific(TAG_BEFORE) &&
                     !opt.isConstructed()) {
                 if (notBefore != null) {
                     throw new CertificateParsingException(
@@ -159,7 +159,7 @@ public class PrivateKeyUsageExtension extends Extension
                 str = new DerInputStream(opt.toByteArray());
                 notBefore = str.getGeneralizedTime();
 
-            } else if (opt.isContextSpecific((byte) TAG_AFTER) &&
+            } else if (opt.isContextSpecific(TAG_AFTER) &&
                        !opt.isConstructed()) {
                 if (notAfter != null) {
                     throw new CertificateParsingException(
